Use Redux Toolkit
Replace reducer with slice
